add feedback from santiago
